requirements.txt
numpy
Pillow
mtcnn
keras-facenet
tensorflow
flask
flask-cors
flask-pymongo
scipy
scikit-learn